Ship the in-memory row-state store for the dry-run diagnostic
The flag-off dry run composes the reconstruction sweep into an in-memory
row-state adapter, but that adapter lived in inmemory.js, which
.dockerignore strips from the production image as a test double. The
diagnostic therefore failed to import it at startup and the server never
listened.

Move the adapter implementation into in-memory-store.js, which ships in
the image, and keep inmemory.js as a re-export for the test imports that
rely on the conventional name. The dry-run diagnostic imports the store
directly; every existing importer is unchanged.
